requirements.txt
CTkMenuBar==0.9
customtkinter==5.2.2
darkdetect==0.8.0
packaging==26.0
pillow==10.4.0
PyYAML==6.0.3
tkinterdnd2==0.4.3
